chore: leverages versions refactor to simplify draft query logic
