+++ package.json
@@ -134,6 +134,7 @@
     "minimist": "^1.2.0",
     "mkdirp": "^1.0.4",
     "mongoose": "6.5.0",
+    "mongoose-aggregate-paginate-v2": "^1.0.6",
     "mongoose-paginate-v2": "^1.6.1",
     "nodemailer": "^6.4.2",
     "object-to-formdata": "^4.1.0",
